Refactor dashboard to unify sections
