+++ frontend/components/dashboard/KnowledgeBaseList.tsx
@@ -1,0 +1,60 @@
+"use client";
+
+import { useEffect, useState } from "react";
+import { useRouter } from "next/navigation";
+import { Button } from "@/components/ui/button";
+import { Pencil, Trash2 } from "lucide-react";
+
+interface Props {
+  token: string | null;
+}
+
+export default function KnowledgeBaseList({ token }: Props) {
+  const router = useRouter();
+  const [data, setData] = useState<any[]>([]);
+  const [loading, setLoading] = useState(true);
+
+  useEffect(() => {
+    if (!token) return;
+    fetch("http://localhost:8000/api/v1/knowledgebase/", {
+      headers: { Authorization: `Bearer ${token}` },
+    })
+      .then((res) => res.json())
+      .then(setData)
+      .finally(() => setLoading(false));
+  }, [token]);
+
+  return (
+    <div className="flex-1 overflow-y-auto px-4 sm:px-8 py-8">
+      {loading ? (
+        <div>Loading...</div>
+      ) : (
+        <table className="min-w-full text-sm border border-gray-200 bg-white rounded-lg">
+          <thead className="bg-gray-50 text-gray-700">
+            <tr>
+              <th className="px-4 py-2 text-left font-semibold">Name</th>
+              <th className="px-4 py-2 text-left font-semibold">Description</th>
+              <th className="px-4 py-2 w-20 text-center font-semibold">Actions</th>
+            </tr>
+          </thead>
+          <tbody>
+            {data.map((kb) => (
+              <tr key={kb.id} className="border-t">
+                <td className="px-4 py-2">{kb.name}</td>
+                <td className="px-4 py-2">{kb.description}</td>
+                <td className="px-4 py-2 flex items-center justify-center gap-2">
+                  <Button variant="ghost" size="icon">
+                    <Pencil className="w-4 h-4" />
+                  </Button>
+                  <Button variant="ghost" size="icon">
+                    <Trash2 className="w-4 h-4" />
+                  </Button>
+                </td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
+      )}
+    </div>
+  );
+}
